Move JH/Eldarion loading logic to separate files
